more code re-arrangement so that vscode lsp dependency to be restricted to language service code.
